splitted _utils.py file in _train_utils.py, _early_stopping.py, _data_utils.py, adjust deep methods for get_trained_autoencoder using FlexibleAutoencoder, removed SimpleAutoencoder #31
