+++ cluspy/deep/__init__.py
@@ -7,7 +7,8 @@ from .enrc import ENRC
 from .simple_autoencoder import Simple_Autoencoder
 from .stacked_autoencoder import Stacked_Autoencoder
 from .flexible_autoencoder import FlexibleAutoencoder
-from ._utils import get_dataloader
+from ._data_utils import get_dataloader
+from ._train_utils import get_trained_autoencoder
 from ._utils import encode_batchwise
 from ._utils import predict_batchwise
 
@@ -23,5 +24,6 @@ __all__ = ['DEC',
            'Stacked_Autoencoder',
            'DipEncoder',
            'get_dataloader',
+           'get_trained_autoencoder',
            'encode_batchwise',
            'predict_batchwise']
