+++ cluspy/deep/dcn.py
@@ -6,8 +6,10 @@ conference on machine learning. PMLR, 2017.
 @authors Lukas Miklautz, Dominik Mautz
 """
 
-from cluspy.deep._utils import detect_device, get_trained_autoencoder, encode_batchwise, \
-    squared_euclidean_distance, predict_batchwise, get_dataloader
+from cluspy.deep._utils import detect_device, encode_batchwise, \
+    squared_euclidean_distance, predict_batchwise
+from ._data_utils import get_dataloader
+from ._train_utils import get_trained_autoencoder 
 import torch
 from sklearn.cluster import KMeans
 from sklearn.base import BaseEstimator, ClusterMixin
@@ -19,11 +21,9 @@ def _dcn(X, n_clusters, batch_size, pretrain_learning_rate, clustering_learning_
     device = detect_device()
     trainloader = get_dataloader(X, batch_size, True, False)
     testloader = get_dataloader(X, batch_size, False, False)
-    if autoencoder is None:
-        autoencoder = get_trained_autoencoder(trainloader, pretrain_learning_rate, pretrain_epochs, device,
-                                              optimizer_class, loss_fn, X.shape[1], embedding_size)
-    else:
-        autoencoder.to(device)
+    
+    autoencoder = get_trained_autoencoder(trainloader, pretrain_learning_rate, pretrain_epochs, device,
+                                          optimizer_class, loss_fn, X.shape[1], embedding_size, autoencoder)
     # Execute kmeans in embedded space
     embedded_data = encode_batchwise(testloader, autoencoder, device)
     kmeans = KMeans(n_clusters=n_clusters)
